refactor(layout): extract table cell helpers
Move numeric and currency cell normalization out of table detection so the remaining layout table code can focus on row grouping and table assembly.
