+++ src/core/layout/tables.ts
@@ -1,5 +1,13 @@
 import type { LayoutLine, LayoutTable } from '../../types/index.js';
 import { type BBox, mode, round2, unionBox } from './geometry.js';
+import {
+  isCurrencyOnlyCell,
+  isTableNumericCell,
+  normalizeTableCurrencyCells,
+  numericColumnMatchRight,
+} from './tableCells.js';
+
+export { isTableNumericCell } from './tableCells.js';
 
 interface IndexedLayoutRow {
   row: LayoutLine[];
@@ -363,16 +371,6 @@ function isAlignedNumericContinuationRow(row: LayoutLine[], columnRights: number
   return matchedCells >= TABLE_ROW_MIN_NUMERIC_CELLS && matchedColumns.size >= TABLE_ROW_MIN_NUMERIC_CELLS;
 }
 
-function numericColumnMatchRight(cell: LayoutLine, nextNumericCell: LayoutLine | undefined): number {
-  const trailing = trailingCurrencyForNextValue(cell.text, nextNumericCell);
-  if (!trailing) return cell.x + cell.width;
-
-  const trimmed = cell.text.trim();
-  const valueText = trimmed.slice(0, -trailing.length).trimEnd();
-  if (trimmed.length === 0 || valueText.length === 0) return cell.x + cell.width;
-  return cell.x + cell.width * (valueText.length / trimmed.length);
-}
-
 function attachLabelContinuationRows(row: LayoutLine[], rowIndex: number, allRows: LayoutLine[][]): LayoutLine[] {
   const label = row.find((line) => !isTableNumericCell(line.text) && /[\p{L}]/u.test(line.text));
   if (!label) return row;
@@ -457,16 +455,6 @@ function hasTableLabelCell(row: LayoutLine[]): boolean {
   return row.some((line) => !isTableNumericCell(line.text) && /[\p{L}]/u.test(line.text));
 }
 
-export function isTableNumericCell(text: string): boolean {
-  const trimmed = text.trim();
-  if (trimmed.length === 0 || !/\d/u.test(trimmed)) return false;
-  const withoutRatioSuffix = trimmed.replace(/(?<=\d)\s*[xX]$/u, '');
-  const withoutScoreWords = withoutRatioSuffix
-    .replace(/\b(?:below|under|over|above|about|approximately)\s+(?=\d)/giu, '')
-    .replace(/(?<=\d)(?:st|nd|rd|th)\b/giu, '');
-  return withoutScoreWords.replace(/[0-9.,()%$¥€£+\-/~\s·⋅∙×^]/gu, '').length === 0;
-}
-
 function rowY(row: LayoutLine[]): number {
   return Math.min(...row.map((line) => line.y));
 }
@@ -495,45 +483,4 @@ function toLayoutTable(rows: LayoutLine[][]): LayoutTable {
       })),
     })),
   };
-}
-
-function normalizeTableCurrencyCells(row: LayoutLine[]): LayoutLine[] {
-  const normalized: LayoutLine[] = [];
-  let pendingCurrency: string | undefined;
-  for (let i = 0; i < row.length; i++) {
-    const cell = row[i];
-    const text = cell.text.trim();
-    if (isCurrencyOnlyCell(text) && isTableNumericCell(row[i + 1]?.text ?? '')) {
-      pendingCurrency = text;
-      continue;
-    }
-
-    const trailing = trailingCurrencyForNextValue(text, row[i + 1]);
-    const textWithoutTrailing = trailing ? text.slice(0, -trailing.length).trimEnd() : text;
-    const nextText = pendingCurrency ? `${pendingCurrency} ${textWithoutTrailing}` : textWithoutTrailing;
-    normalized.push({ ...cell, text: nextText });
-    pendingCurrency = trailing;
-  }
-  if (pendingCurrency) {
-    normalized.push({
-      text: pendingCurrency,
-      x: row.at(-1)?.x ?? 0,
-      y: rowY(row),
-      width: 0,
-      height: rowBottom(row) - rowY(row),
-      fontSize: row.at(-1)?.fontSize ?? 0,
-    });
-  }
-  return normalized;
-}
-
-function isCurrencyOnlyCell(text: string): boolean {
-  return /^[$¥€£]$/u.test(text.trim());
-}
-
-function trailingCurrencyForNextValue(text: string, next: LayoutLine | undefined): string | undefined {
-  if (!next || !isTableNumericCell(next.text)) return undefined;
-  const match = /^(.+?)\s*([$¥€£])$/u.exec(text.trim());
-  if (!match) return undefined;
-  return isTableNumericCell(match[1]) ? match[2] : undefined;
 }
